chore: move to auth directory
