+++ src/main/java/com/codeplanks/home360/domain/auth/PasswordResetRequest.java
@@ -1,4 +1,5 @@
-package com.codeplanks.home360.domain.passwordReset;
+/* (C)2024 */
+package com.codeplanks.home360.domain.auth;
 
 import com.codeplanks.home360.validation.Password;
 import jakarta.validation.constraints.Email;
@@ -12,7 +13,7 @@ public class PasswordResetRequest {
   @Email(message = "Enter a valid email address")
   private String email;
 
-  private String oldPassword;
+  //  private String oldPassword;
 
   @Password(message = "Password is required")
   private String newPassword;
