--- src/main/java/com/codeplanks/home360/exception/RefreshTokenExpiredException.java
@@ -1,19 +1,0 @@
-package com.codeplanks.home360.exception;
-
-import org.springframework.http.HttpStatus;
-import org.springframework.web.bind.annotation.ResponseStatus;
-
-@ResponseStatus(HttpStatus.UNAUTHORIZED)
-public class RefreshTokenExpiredException extends RuntimeException {
-  public RefreshTokenExpiredException(String message) {
-    super(message);
-  }
-
-  public RefreshTokenExpiredException(String message, Throwable cause) {
-    super(message, cause);
-  }
-
-  public RefreshTokenExpiredException(Throwable cause) {
-    super(cause);
-  }
-}
